requirements.txt
absl-py==0.1.10
apturl==0.5.2
asn1crypto==0.22.0
bleach==1.5.0
Brlapi==0.6.5
certifi==2018.1.18
chardet==3.0.4
click==6.7
command-not-found==0.3
cryptography==1.9
cupshelpers==1.0
cycler==0.10.0
defer==1.0.6
distro-info==0.17
Flask==0.12.2
Flask-Cors==3.0.3
html5lib==0.9999999
httplib2==0.9.2
idna==2.6
itsdangerous==0.24
Jinja2==2.10
keyring==10.4.0
keyrings.alt==2.2
language-selector==0.1
launchpadlib==1.10.5
lazr.restfulclient==0.13.5
lazr.uri==1.0.3
louis==3.0.0
Mako==1.0.7
Markdown==2.6.11
MarkupSafe==1.0
matplotlib==2.1.2
numpy==1.14.1
oauth==1.0.1
oauthlib==2.0.6
olefile==0.44
pexpect==4.2.1
Pillow==4.1.1
protobuf==3.5.1
pycrypto==2.6.1
pycups==1.9.73
pygobject==3.24.1
pymongo==3.6.0
pyparsing==2.2.0
python-apt==1.4.0b3
python-dateutil==2.6.1
python-debian==0.1.30
pytz==2018.3
pyxdg==0.25
PyYAML==3.12
reportlab==3.4.0
requests==2.18.4
requests-oauthlib==0.8.0
scipy==1.0.0
screen-resolution-extra==0.0.0
SecretStorage==2.3.1
simplejson==3.11.1
six==1.11.0
system-service==0.3
systemd-python==234
tensorflow-gpu==1.5.0
tensorflow-tensorboard==1.5.1
tweepy==3.5.0
ubuntu-drivers-common==0.0.0
ufw==0.35
unattended-upgrades==0.1
urllib3==1.22
usb-creator==0.3.3
wadllib==1.3.2
Werkzeug==0.14.1
xkit==0.0.0
zope.interface==4.3.2
